+++ python/mead/__init__.py
@@ -1,2 +1,3 @@
 from mead.tasks import *
 from mead.utils import *
+from mead.exporters import *
